Move ContinuousExecutionGate to deployment.internal
