+++ subprojects/core/src/main/java/org/gradle/deployment/internal/ContinuousExecutionGate.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.initialization;
+package org.gradle.deployment.internal;
 
 import javax.annotation.Nonnull;
