+++ subprojects/core/src/main/java/org/gradle/deployment/internal/DefaultContinuousExecutionGate.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.initialization;
+package org.gradle.deployment.internal;
 
 import com.google.common.collect.Lists;
 import org.gradle.internal.UncheckedException;
