+++ subprojects/launcher/src/test/groovy/org/gradle/tooling/internal/provider/ContinuousBuildActionExecutorTest.groovy
@@ -18,13 +18,13 @@ package org.gradle.tooling.internal.provider
 
 import org.gradle.api.internal.StartParameterInternal
 import org.gradle.api.internal.file.TestFiles
+import org.gradle.deployment.internal.DefaultContinuousExecutionGate
 import org.gradle.deployment.internal.Deployment
 import org.gradle.deployment.internal.DeploymentInternal
 import org.gradle.deployment.internal.DeploymentRegistryInternal
 import org.gradle.initialization.BuildRequestMetaData
 import org.gradle.initialization.DefaultBuildCancellationToken
 import org.gradle.initialization.DefaultBuildRequestContext
-import org.gradle.initialization.DefaultContinuousExecutionGate
 import org.gradle.initialization.NoOpBuildEventConsumer
 import org.gradle.integtests.fixtures.RedirectStdIn
 import org.gradle.internal.buildevents.BuildStartedTime
